Default all int entries in the dhcp header to unsigned.

+++ dhcp.c
@@ -124,7 +124,7 @@ size_t send_message (const interface_t *iface, const dhcp_t *dhcp,
 		case ARPHRD_IEEE1394:
 		case ARPHRD_INFINIBAND:
 			if (message->ciaddr == 0)
-				message->flags = (int16_t) htons (BROADCAST_FLAG);
+				message->flags = htons (BROADCAST_FLAG);
 			message->hwlen = 0;
 			break;
 		default:
